refactor: adopt vterm
There's a lot of unicode-related code in libvterm which is duplicated by
the functionality in mbyte.c and utf8proc. To ensure a consistent
experience we have to adopt the code sections we're already
implementing.

Working off the changes from termkey PR because I can't be arsed to
surgically separate changes. Merge that PR first.
