+++ src/nvim/os/win_defs.h
@@ -16,9 +16,6 @@
 #include <sys/stat.h>
 #include <windows.h>
 
-// vterm.h defines an `unsigned int small` in a struct, triggering error C2632
-#undef small
-
 // Windows does not have S_IFLNK but libuv defines it
 // and sets the flag for us when calling uv_fs_stat.
 #include <uv.h>
