+++ src/nvim/tui/driver-csi.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "nvim/tui/termkey_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "tui/driver-csi.h.generated.h"
+#endif
