+++ src/nvim/tui/driver-ti.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "nvim/tui/termkey_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "tui/driver-ti.h.generated.h"
+#endif
