+++ src/nvim/tui/termkey-internal.h
@@ -1,31 +1,29 @@
-#ifndef GUARD_TERMKEY_INTERNAL_H_
-#define GUARD_TERMKEY_INTERNAL_H_
+#pragma once
+
+#include <stdint.h>
+
+#include "nvim/tui/termkey_defs.h"
 
 #define HAVE_TERMIOS
-
 #ifdef _WIN32
 # undef HAVE_TERMIOS
 #endif
 
-#include "termkey.h"
-
-#include <stdint.h>
 #ifdef HAVE_TERMIOS
 # include <termios.h>
 #endif
 
 #ifdef _MSC_VER
-#include <BaseTsd.h>
+# include <BaseTsd.h>
 typedef SSIZE_T ssize_t;
 #endif
 
-struct TermKeyDriver
-{
-  const char      *name;
-  void          *(*new_driver)(TermKey *tk, const char *term);
-  void           (*free_driver)(void *info);
-  int            (*start_driver)(TermKey *tk, void *info);
-  int            (*stop_driver)(TermKey *tk, void *info);
+struct TermKeyDriver {
+  const char *name;
+  void *(*new_driver)(TermKey *tk, const char *term);
+  void (*free_driver)(void *info);
+  int (*start_driver)(TermKey *tk, void *info);
+  int (*stop_driver)(TermKey *tk, void *info);
   TermKeyResult (*peekkey)(TermKey *tk, void *info, TermKeyKey *key, int force, size_t *nbytes);
 };
 
@@ -38,21 +36,21 @@ struct keyinfo {
 
 struct TermKeyDriverNode;
 struct TermKeyDriverNode {
-  struct TermKeyDriver     *driver;
-  void                      *info;
+  struct TermKeyDriver *driver;
+  void *info;
   struct TermKeyDriverNode *next;
 };
 
 struct TermKey {
-  int    fd;
-  int    flags;
-  int    canonflags;
+  int fd;
+  int flags;
+  int canonflags;
   unsigned char *buffer;
-  size_t buffstart; // First offset in buffer
-  size_t buffcount; // NUMBER of entires valid in buffer
-  size_t buffsize; // Total malloc'ed size
-  size_t hightide; /* Position beyond buffstart at which peekkey() should next start
-                    * normally 0, but see also termkey_interpret_csi */
+  size_t buffstart;  // First offset in buffer
+  size_t buffcount;  // NUMBER of entires valid in buffer
+  size_t buffsize;   // Total malloc'ed size
+  size_t hightide;   // Position beyond buffstart at which peekkey() should next start
+                     // normally 0, but see also termkey_interpret_csi
 
 #ifdef HAVE_TERMIOS
   struct termios restore_termios;
@@ -62,12 +60,12 @@ struct TermKey {
   TermKey_Terminfo_Getstr_Hook *ti_getstr_hook;
   void *ti_getstr_hook_data;
 
-  int waittime; // msec
+  int waittime;  // msec
 
-  char   is_closed;
-  char   is_started;
+  char is_closed;
+  char is_started;
 
-  int  nkeynames;
+  int nkeynames;
   const char **keynames;
 
   // There are 32 C0 codes
@@ -78,7 +76,7 @@ struct TermKey {
   // Now some "protected" methods for the driver to call but which we don't
   // want exported as real symbols in the library
   struct {
-    void (*emit_codepoint)(TermKey *tk, long codepoint, TermKeyKey *key);
+    void (*emit_codepoint)(TermKey *tk, int codepoint, TermKeyKey *key);
     TermKeyResult (*peekkey_simple)(TermKey *tk, TermKeyKey *key, int force, size_t *nbytes);
     TermKeyResult (*peekkey_mouse)(TermKey *tk, TermKeyKey *key, size_t *nbytes);
   } method;
@@ -86,27 +84,26 @@ struct TermKey {
 
 static inline void termkey_key_get_linecol(const TermKeyKey *key, int *line, int *col)
 {
-  if(col)
-    *col  = (unsigned char)key->code.mouse[1] | ((unsigned char)key->code.mouse[3] & 0x0f) << 8;
+  if (col) {
+    *col = (unsigned char)key->code.mouse[1] | ((unsigned char)key->code.mouse[3] & 0x0f) << 8;
+  }
 
-  if(line)
+  if (line) {
     *line = (unsigned char)key->code.mouse[2] | ((unsigned char)key->code.mouse[3] & 0x70) << 4;
+  }
 }
 
 static inline void termkey_key_set_linecol(TermKeyKey *key, int line, int col)
 {
-  if(line > 0xfff)
+  if (line > 0xfff) {
     line = 0xfff;
+  }
 
-  if(col > 0x7ff)
+  if (col > 0x7ff) {
     col = 0x7ff;
+  }
 
-  key->code.mouse[1] = (line & 0x0ff);
-  key->code.mouse[2] = (col & 0x0ff);
+  key->code.mouse[1] = (char)(line & 0x0ff);
+  key->code.mouse[2] = (char)(col & 0x0ff);
   key->code.mouse[3] = (line & 0xf00) >> 8 | (col & 0x300) >> 4;
 }
-
-extern struct TermKeyDriver termkey_driver_csi;
-extern struct TermKeyDriver termkey_driver_ti;
-
-#endif
